Fix/bug

+++ src/utils/product.utils.js
@@ -13,30 +13,52 @@ const getAllProduct = async (page, limit) => {
 };
 
 const getProductById = async (productId) => {
-  const product = await ProductModel.findById(productId);
+  const product = await ProductModel.findById(productId).lean();
+  return product;
+};
+
+const getProductByName = async (productName) => {
+  const product = await ProductModel.findOne({ name: productName });
   return product;
 };
 
 const updateProductById = async (productData) => {
-  const product = await ProductModel.findByIdAndUpdate(
+  const product = await getProductById(productData.productId);
+  if (!product) {
+    console.log("product not found");
+    throw new Error("RESOURCE_NOT_FOUND");
+  }
+  const updatedProduct = await ProductModel.findByIdAndUpdate(
     productData.productId,
     {
       ...productData,
     },
     { new: true }
   );
-  return product;
+  return updatedProduct;
 };
 
 const deleteProductById = async (productId) => {
+  const product = await getProductById(productId);
+  if (!product) {
+    console.log("product not found");
+    throw new Error("NOT_FOUND");
+  }
   await ProductModel.findByIdAndDelete(productId);
   return "product deleted successsfully !";
+};
+
+const countDocuments = async (query) => {
+  const documents = await ProductModel.countDocuments(query);
+  return documents;
 };
 
 export default {
   saveProduct,
   getAllProduct,
   getProductById,
+  getProductByName,
   updateProductById,
   deleteProductById,
+  countDocuments,
 };
